backend/requirements.txt
# FastAPI and web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database (PostgreSQL)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Data Science and AI (Windows-compatible versions)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image processing
Pillow>=10.4.0
opencv-python>=4.8.0

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0 